"check_style_score -> 5 of 5 passed. style score: 0/1 "

+++ m2/Assignment-3/Solution.java
@@ -8,13 +8,13 @@ public class Solution {
 		Scanner s=new Scanner(System.in);      
         int base = s.nextInt();
         int exponent = s.nextInt();
-        int result=power(base, exponent);
+        long result=power(base, exponent);
         System.out.println(result);
 	}
 	/*
 	Need to write the power function and print the output.
 	*/
-	static int power(int b, int e){
+	static long power(int b, int e){
 		if (e != 0){
 			return (b*(power(b, e-1)));
 		}
